fix checking snapshots from earlier upgrades
Previously this check was done by searching for snapshots *globally*,
which is a bad idea if you e.g. have backups on another pool on the
same system.
This patche ensures iocell is only looking at its own datasets, which
also speeds up that check considerably on systems with multiple large
pools.